requirements.txt
Flask
tensorflow
numpy
Pillow
flask-cors
tensorflow_hub
openai